+++ app/src/main/java/com/m/androidNativeApp/MainActivity.java
@@ -23,7 +23,7 @@ import com.apollographql.apollo.api.Response;
 import com.apollographql.apollo.exception.ApolloException;
 import com.apollographql.apollo.fetcher.ApolloResponseFetchers;
 import com.apollographql.apollo.fetcher.ResponseFetcher;
-import com.m.helper.Client;
+import com.m.services.dataSync.Client;
 import com.m.helper.CreateTask;
 import com.m.helper.Item;
 import com.m.helper.ItemAdapter;
@@ -42,9 +42,9 @@ import org.jetbrains.annotations.NotNull;
 
 import java.util.ArrayList;
 
-import static com.m.helper.LoginActivity.RE_AUTH;
-import static com.m.helper.LoginActivity.mAuthStateManager;
-import static com.m.helper.LoginActivity.mobileService;
+import static com.m.services.appAuth.AppAuthActivity.RE_AUTH;
+import static com.m.services.appAuth.AppAuthActivity.mAuthStateManager;
+import static com.m.services.appAuth.AppAuthActivity.mobileService;
 
 
 
@@ -276,7 +276,7 @@ public class MainActivity extends AppCompatActivity implements MessageHandler {
 
     public void reAuthorise() {
         RE_AUTH = 403;
-        Intent redirectToRefreshToken = new Intent(MainActivity.this, LoginActivity.class);
+        Intent redirectToRefreshToken = new Intent(MainActivity.this, AppAuthActivity.class);
         startActivity(redirectToRefreshToken);
     }
