+++ app/src/main/java/com/m/services/appAuth/AuthStateManager.java
@@ -1,4 +1,4 @@
-package com.m.helper.Auth;
+package com.m.services.appAuth;
 
 
 import android.content.Context;
