+++ app/src/main/java/com/m/services/dataSync/Client.java
@@ -1,4 +1,4 @@
-package com.m.helper;
+package com.m.services.dataSync;
 
 import android.content.Context;
 
@@ -25,9 +25,24 @@ import okhttp3.Request;
 
 public class Client {
 
+    /**
+     * Setting up Apollo Client
+     * @param serverUrl
+     *           Server URL provided in mobile-service.json, in our example passed in as a parameter
+     *           in MainActivity to construct Apollo Client.
+     * @param authHeader
+     *           Token received.
+     * @param context
+     *          Context required for normalized cache.
+     * @return
+     *          Returns interceptor chain from OkHttp
+     */
     public static ApolloClient setupApollo(String serverUrl, String authHeader, Context context) {
 
 
+        /**
+         * Resolver that can be used to access records from cache
+         */
         CacheKeyResolver resolver = new CacheKeyResolver() {
             @NotNull
             @Override
@@ -51,11 +66,20 @@ public class Client {
         };
 
         String DB_CACHE_NAME = "myapp";
+
+
+        /**
+         *  Creating normalized cache for our app to enable offline support for our application
+         */
         ApolloSqlHelper apolloSqlHelper = new ApolloSqlHelper(context, DB_CACHE_NAME);
 
         NormalizedCacheFactory<LruNormalizedCache> cacheFactory = new LruNormalizedCacheFactory(EvictionPolicy.NO_EVICTION)
                 .chain(new SqlNormalizedCacheFactory(apolloSqlHelper));
 
+
+        /**
+         * OkHttp is handling network requests for our client.
+         */
         OkHttpClient okHttpClient = new OkHttpClient
                 .Builder()
                 .addInterceptor(chain -> {
@@ -66,9 +90,18 @@ public class Client {
                 })
                 .build();
 
+
+        /**
+         * Setting up authorization to support subscriptions
+         */
         Map<String, Object> connectionParams = new HashMap<>();
         connectionParams.put("Authorization", authHeader);
 
+
+        /**
+         * Building Apollo Client, passing in serverUrl, normalized cache, okHttp client and setting up
+         * subscription mechanism.
+         */
         return ApolloClient.builder()
                 .enableAutoPersistedQueries(true)
                 .serverUrl(serverUrl)
@@ -77,5 +110,6 @@ public class Client {
                 .subscriptionTransportFactory(new WebSocketSubscriptionTransport.Factory(serverUrl, okHttpClient))
                 .okHttpClient(okHttpClient)
                 .build();
+
     }
 }
